Padronizando a tipagem para Integer

+++ slack_hafutech/hafutech/src/main/java/sptech/school/hafutech/repository/EmpresaRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 import sptech.school.hafutech.entity.Empresa;
 
 @Repository
-public interface EmpresaRepository extends JpaRepository<Empresa, Long> {
+public interface EmpresaRepository extends JpaRepository<Empresa, Integer> {
     @Query("SELECT id FROM Empresa ORDER BY id DESC LIMIT 1")
-    Long pegarUltimoId();
+    Integer pegarUltimoId();
 
 }
